+++ src/execute/eval.c
@@ -2,8 +2,8 @@
 
 ErrorCode eval_expr(Env env, Expr expr, struct Term* result) {
     // DEBUG
-    // char buf[1024]; print_expr(expr, env->dict, buf);
-    // debug(2, "eval_expr(%s) - ", buf);
+    char buf[1024]; print_expr(expr, env->dict, buf);
+    debug(2, "eval_expr(%s) - ", buf);
     // print_dict(env->dict, buf);
     // debug(2, "%s\n\n", buf);
     
@@ -45,14 +45,12 @@ ErrorCode eval_expr(Env env, Expr expr, struct Term* result) {
 // Evaluate a combination (a list containing at least one element)
 // Here, `expr` does not contain the opening parenthesis
 ErrorCode eval_combination(Env env, Expr expr, struct Term* result) {
-    debug(2, "eval_combination - ");
     // At this point, there will be no exact match, the longest matching
     // subexpression will be at least one element shorter than the expression
     // given as parameter.
     size_t matching_bytes;
     struct Term* match = find_longest_match(env, expr, &matching_bytes);
     if (match == NULL) {
-        debug(2, "No match\n");
         // No match, evaluate expression from left to right. The fact that not
         // even the first element matches is not a problem, since it could be an
         // uncached list. Therefore we still need to try to evaluate it.
@@ -76,7 +74,6 @@ ErrorCode eval_combination(Env env, Expr expr, struct Term* result) {
         }
         return SUCCESS;
     } else {
-        debug(2, "Match\n");
         // There is a match, take its value and continue evaluating the rest of
         // the expression from left to right
         return apply(env, *match, expr + matching_bytes, result);
@@ -84,7 +81,6 @@ ErrorCode eval_combination(Env env, Expr expr, struct Term* result) {
 }
 
 ErrorCode apply(Env env, struct Term t, Expr e, struct Term* result) {
-    debug(2, "apply - ");
     if (t.type == ValTerm) {
         return ERROR;
     }
@@ -103,11 +99,11 @@ ErrorCode apply(Env env, struct Term t, Expr e, struct Term* result) {
                 if (is_equal_expr(t.expr, evaled.expr)) {
                     return ERROR;
                 }
-            }            
+            }
         }
+        // free_term(evaled); // seems to do nothing?
         return apply(env, t, e, result);
     } else {
-        debug(2, "abstraction - calling apply\n");
         return t.abs.apply(env, e, t.abs.closure, result);
     }
 }